overlay/app/requirements.txt
uvicorn~=0.15.0
fastapi~=0.70.0
python-dotenv~=0.19.2